services/ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
easyocr==1.7.0
redis==5.0.1
Pillow==9.5.0
numpy==1.24.3
PyMuPDF==1.23.8